pyproject.toml
[tool.poetry]
name = "twitter-b2bit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "twitter_b2bit"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
pillow = "^9.5.0"
django-stubs = "^4.2.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.7.0"
pyright = "^1.1.309"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
